Created the Check class and created the jumper so that he displays properly
I created the check_guess.py with the Check class which displays the word completion and compares the given guess to the secret word to see if it is correct or not. I also created the jumper as part of a list so that he can be displayed in stages according to how many correct or incorrect guesses the user provides. I added comments throughout the code that explain my thought process somewhat